requirements.txt
astunparse==1.6.3
async-timeout==4.0.2
blessed==1.20.0
cachetools==5.5.0
certifi==2022.9.24
charset-normalizer==2.1.1
cycler==0.11.0
dm-tree==0.1.8
docopt==0.6.2
einops==0.6.1
filelock==3.8.0
fonttools==4.38.0
gast==0.6.0
gpustat==1.1.1
greenlet==3.0.3
gunicorn==21.2.0
h5py==3.8.0
huggingface-hub==0.11.0
idna==3.4
imageio==2.22.4
importlib-metadata==5.1.0
itsdangerous==2.1.2
Jinja2==3.1.3
kiwisolver==1.4.5
MarkupSafe==2.1.5
matplotlib==3.5.3
mkl-fft==1.3.1
mkl-random==1.2.2
mkl-service==2.4.0
networkx==2.6.3
numpy
olefile
opencv-contrib-python==4.5.4.60
opencv-python==4.6.0.66
packaging==21.3
pandas==1.3.5
Pillow
pipreqs==0.4.13
psutil==6.0.0
pyparsing==3.0.9
python-dateutil==2.8.2
pytz==2023.3.post1
PyWavelets==1.3.0
PyYAML==5.4.1
redis==4.5.5
requests==2.28.1
scikit-image==0.19.3
scipy==1.7.3
SQLAlchemy==2.0.28
tabulate==0.9.0
termcolor==2.3.0
thop==0.1.1.post2209072238
tifffile==2021.11.2
timm==0.6.12
torch==1.10.1
torchaudio==0.10.1
torchpack==0.2.2
torchstat==0.0.7
torchsummary==1.5.1
torchvideotransforms @ git+https://github.com/hassony2/torch_videovision@785a810d1df2ccab6901fb98b38e358433e7b17a
torchvision==0.11.2
